requirements.txt
pip>=23.0.0
streamlit>=1.0.0
pillow>=9.3.0